Original change: https://googleplex-android-review.googlesource.com/c/platform/packages/modules/AdServices/+/22362852

+++ adservices/service-core/java/com/android/adservices/data/measurement/IMeasurementDao.java
@@ -40,21 +40,15 @@ import java.util.Collection;
 import java.util.List;
 import java.util.Set;
 
-/**
- * Interface for Measurement related data access operations.
- */
+/** Interface for Measurement related data access operations. */
 public interface IMeasurementDao {
-    /**
-     * Set the transaction.
-     */
+    /** Set the transaction. */
     void setTransaction(ITransaction transaction);
 
     /** Add an entry to the Trigger datastore. */
     void insertTrigger(Trigger trigger) throws DatastoreException;
 
-    /**
-     * Returns list of ids for all pending {@link Trigger}.
-     */
+    /** Returns list of ids for all pending {@link Trigger}. */
     List<String> getPendingTriggerIds() throws DatastoreException;
 
     /**
@@ -134,18 +128,27 @@ public interface IMeasurementDao {
      * Gets the count of distinct Uri's of destinations in the Source table in a time window with
      * matching publisher, enrollment, and ACTIVE status, excluding a given destination.
      */
-    Integer countDistinctDestinationsPerPublisherXEnrollmentInActiveSource(Uri publisher,
-            @EventSurfaceType int publisherType, String enrollmentId,
-            List<Uri> excludedDestinations, @EventSurfaceType int destinationType,
-            long windowStartTime, long windowEndTime) throws DatastoreException;
+    Integer countDistinctDestinationsPerPublisherXEnrollmentInActiveSource(
+            Uri publisher,
+            @EventSurfaceType int publisherType,
+            String enrollmentId,
+            List<Uri> excludedDestinations,
+            @EventSurfaceType int destinationType,
+            long windowStartTime,
+            long windowEndTime)
+            throws DatastoreException;
 
     /**
      * Gets the count of distinct IDs of enrollments in the Source table in a time window with
      * matching publisher and destination, excluding a given enrollment ID.
      */
-    Integer countDistinctEnrollmentsPerPublisherXDestinationInSource(Uri publisher,
-            @EventSurfaceType int publisherType, List<Uri> destinations,
-            String excludedEnrollmentId, long windowStartTime, long windowEndTime)
+    Integer countDistinctEnrollmentsPerPublisherXDestinationInSource(
+            Uri publisher,
+            @EventSurfaceType int publisherType,
+            List<Uri> destinations,
+            String excludedEnrollmentId,
+            long windowStartTime,
+            long windowEndTime)
             throws DatastoreException;
 
     /**
@@ -221,12 +224,12 @@ public interface IMeasurementDao {
 
     /**
      * Queries and returns the {@link AggregateReport}
+     *
      * @param aggregateReportId Id of the request Aggregate Report
      * @return the request Aggregate Report; Null in case of SQL failure
      */
     @Nullable
-    AggregateReport getAggregateReport(String aggregateReportId)
-            throws DatastoreException;
+    AggregateReport getAggregateReport(String aggregateReportId) throws DatastoreException;
 
     /**
      * Queries and returns the {@link DebugReport}
@@ -269,14 +272,10 @@ public interface IMeasurementDao {
      */
     void markAggregateDebugReportDelivered(String aggregateReportId) throws DatastoreException;
 
-    /**
-     * Saves the {@link EventReport} to datastore.
-     */
+    /** Saves the {@link EventReport} to datastore. */
     void insertEventReport(EventReport eventReport) throws DatastoreException;
 
-    /**
-     * Deletes the {@link EventReport} from datastore.
-     */
+    /** Deletes the {@link EventReport} from datastore. */
     void deleteEventReport(EventReport eventReport) throws DatastoreException;
 
     /** Deletes the {@link DebugReport} from datastore. */
@@ -291,9 +290,7 @@ public interface IMeasurementDao {
     /** Returns list of all debug event reports. */
     List<String> getPendingDebugEventReportIds() throws DatastoreException;
 
-    /**
-     * Returns list of all pending event reports for a given app right away.
-     */
+    /** Returns list of all pending event reports for a given app right away. */
     List<String> getPendingEventReportIdsForGivenApp(Uri appName) throws DatastoreException;
 
     /**
@@ -322,7 +319,7 @@ public interface IMeasurementDao {
     /**
      * Mark relevant source as install attributed.
      *
-     * @param uri            package identifier
+     * @param uri package identifier
      * @param eventTimestamp timestamp of installation event
      */
     void doInstallAttribution(Uri uri, long eventTimestamp) throws DatastoreException;
@@ -330,13 +327,11 @@ public interface IMeasurementDao {
     /**
      * Undo any install attributed source events.
      *
-     * @param uri            package identifier
+     * @param uri package identifier
      */
     void undoInstallAttribution(Uri uri) throws DatastoreException;
 
-    /**
-     * Save aggregate encryption key to datastore.
-     */
+    /** Save aggregate encryption key to datastore. */
     void insertAggregateEncryptionKey(AggregateEncryptionKey aggregateEncryptionKey)
             throws DatastoreException;
 
@@ -347,14 +342,10 @@ public interface IMeasurementDao {
     List<AggregateEncryptionKey> getNonExpiredAggregateEncryptionKeys(long expiry)
             throws DatastoreException;
 
-    /**
-     *  Remove aggregate encryption keys from the datastore older than {@code expiry}.
-     */
+    /** Remove aggregate encryption keys from the datastore older than {@code expiry}. */
     void deleteExpiredAggregateEncryptionKeys(long expiry) throws DatastoreException;
 
-    /**
-     * Save unencrypted aggregate payload to datastore.
-     */
+    /** Save unencrypted aggregate payload to datastore. */
     void insertAggregateReport(AggregateReport payload) throws DatastoreException;
 
     /** Save debug report payload to datastore. */
@@ -373,9 +364,7 @@ public interface IMeasurementDao {
     /** Returns list of all debug reports. */
     List<String> getDebugReportIds() throws DatastoreException;
 
-    /**
-     * Returns list of all pending aggregate reports for a given app right away.
-     */
+    /** Returns list of all pending aggregate reports for a given app right away. */
     List<String> getPendingAggregateReportIdsForGivenApp(Uri appName) throws DatastoreException;
 
     /**
@@ -433,7 +422,7 @@ public interface IMeasurementDao {
      * @param retryLimit a long that is used for determining the next valid record to be serviced
      * @param failedOrigins set of origins that have failed during the current run
      */
-    AsyncRegistration fetchNextQueuedAsyncRegistration(short retryLimit, Set<Uri> failedOrigins)
+    AsyncRegistration fetchNextQueuedAsyncRegistration(int retryLimit, Set<Uri> failedOrigins)
             throws DatastoreException;
 
     /**
